+++ src/peer_to_peer_dht.c
@@ -18,7 +18,9 @@
 #include <opendht/opendht_c.h>
 #include "conf.h"
 #include "utils.h"
+#include "jansson.h"
 #include "json_logger.h"
+#include "database.h"
 
 #define DHT_PORT 4222
 #define DHT_BOOTSTRAP_NODE "bootstrap.sentrypeer.org"
@@ -42,7 +44,228 @@ static bool dht_value_callback(const dht_value *value, bool expired,
 	if (config->debug_mode || config->verbose_mode) {
 		fprintf(stderr, "Value callback %s: %.*s\n",
 			expired ? "expired" : "new", (int)data.size, data.data);
-		fprintf(stderr, "TODO: Save into into sqlite\n");
+	}
+
+	if (expired) {
+		return true;
+	}
+
+	if (data.size == 0) {
+		return true;
+	}
+
+	// TODO: Move to a validator function
+	if (data.size > 0) {
+		char *received_value_str = malloc(data.size + 1);
+		memcpy(received_value_str, data.data, data.size);
+		received_value_str[data.size] = '\0';
+
+		// Turn back into JSON
+		json_error_t error;
+		json_t *json = json_loads(received_value_str, 0, &error);
+
+		if (!json_is_object(json)) {
+			if (config->debug_mode || config->verbose_mode) {
+				fprintf(stderr,
+					"JSON from DHT is not valid.\n");
+				fprintf(stderr, "Failed to parse: %s\n",
+					error.text);
+			}
+			json_decref(json);
+			free(received_value_str);
+			return true;
+		}
+
+		json_t *node_id = json_object_get(json, "created_by_node_id");
+		if (!node_id) {
+			if (config->debug_mode || config->verbose_mode) {
+				fprintf(stderr,
+					"JSON from DHT is not valid.\n");
+				fprintf(stderr, "No node_id in JSON.\n");
+			}
+			json_decref(json);
+			free(received_value_str);
+			return true;
+		}
+
+		const char *node_id_str = json_string_value(node_id);
+		if (!node_id_str) {
+			if (config->debug_mode || config->verbose_mode) {
+				fprintf(stderr,
+					"JSON from DHT is not valid.\n");
+				fprintf(stderr, "Node ID is not a string.\n");
+			}
+			json_decref(json);
+			free(received_value_str);
+			return true;
+		}
+
+		uuid_t node_id_uuid_check;
+		if (uuid_parse(node_id_str, node_id_uuid_check) !=
+		    EXIT_SUCCESS) {
+			if (config->debug_mode || config->verbose_mode) {
+				fprintf(stderr,
+					"Node ID uuid in JSON from DHT is not valid.\n");
+			}
+			json_decref(json);
+			free(received_value_str);
+			return true;
+		}
+
+		if (config->debug_mode || config->verbose_mode) {
+			fprintf(stderr, "Node ID from DHT value is: %s\n",
+				node_id_str);
+		}
+
+		// Check it's not from us
+		if (strncmp(node_id_str, config->node_id,
+			    strlen(config->node_id)) == 0) {
+			if (config->debug_mode || config->verbose_mode) {
+				fprintf(stderr,
+					"Node ID from DHT value is the same as ours. Not saving bad_actor.\n");
+			}
+			json_decref(json);
+			free(received_value_str);
+			return true;
+		}
+
+		// TODO: Validate event_uuid and check our database to see if we already have this bad_actor
+		// by using the event_uuid
+		json_t *event_uuid = json_object_get(json, "event_uuid");
+		if (!event_uuid) {
+			if (config->debug_mode || config->verbose_mode) {
+				fprintf(stderr,
+					"JSON from DHT is not valid.\n");
+				fprintf(stderr, "No event_uuid in JSON.\n");
+			}
+			json_decref(json);
+			free(received_value_str);
+			return true;
+		}
+
+		const char *event_uuid_str = json_string_value(event_uuid);
+		if (!event_uuid_str) {
+			if (config->debug_mode || config->verbose_mode) {
+				fprintf(stderr,
+					"JSON from DHT is not valid.\n");
+				fprintf(stderr,
+					"event_uuid is not a string.\n");
+			}
+			json_decref(json);
+			free(received_value_str);
+			return true;
+		}
+
+		if (!is_valid_uuid(event_uuid_str)) {
+			if (config->debug_mode || config->verbose_mode) {
+				fprintf(stderr,
+					"event_uuid in JSON from DHT is invalid.\n");
+			}
+			json_decref(json);
+			free(received_value_str);
+			return true;
+		}
+
+		if (config->debug_mode || config->verbose_mode) {
+			fprintf(stderr, "event_uuid from DHT value is: %s\n",
+				node_id_str);
+		}
+
+		// Check rest of bad_actor keys
+		// TODO: Use bad_actor type instead of hard coding here
+		const char *bad_actor_keys[] = {
+			"event_timestamp", "collected_method", "sip_message",
+			"source_ip",	   "destination_ip",   "called_number",
+			"method",	   "transport_type"
+		};
+
+		int loop_control = 0;
+		while (loop_control <= 7) {
+			json_t *json_value = json_object_get(
+				json, bad_actor_keys[loop_control]);
+			if (!json_value) {
+				if (config->debug_mode ||
+				    config->verbose_mode) {
+					fprintf(stderr,
+						"JSON from DHT is not valid.\n");
+					fprintf(stderr,
+						"No '%s' key in JSON.\n",
+						bad_actor_keys[loop_control]);
+				}
+				json_decref(json);
+				free(received_value_str);
+				return true;
+			}
+		}
+
+		if (!db_bad_actor_exists(event_uuid_str, config)) {
+			// TODO: Move this to a dht_save_bad_actor function
+			// It's not from us, so it's a new bad_actor we want to save
+			if (config->debug_mode || config->verbose_mode) {
+				fprintf(stderr,
+					"Saving new bad_actor from node_id: %s\n",
+					node_id_str);
+			}
+
+			bad_actor *bad_actor_event = bad_actor_new(
+				util_duplicate_string(json_string_value(
+					json_object_get(json, "sip_message"))),
+				util_duplicate_string(json_string_value(
+					json_object_get(json, "source_ip"))),
+				util_duplicate_string(json_string_value(
+					json_object_get(json,
+							"destination_ip"))),
+				util_duplicate_string(json_string_value(
+					json_object_get(json, "called_number"))),
+				util_duplicate_string(json_string_value(
+					json_object_get(json, "method"))),
+				util_duplicate_string(json_string_value(
+					json_object_get(json,
+							"transport_type"))),
+				util_duplicate_string(json_string_value(
+					json_object_get(json, "user_agent"))),
+				util_duplicate_string(json_string_value(
+					json_object_get(json,
+							"collected_method"))),
+				(char *)node_id_str);
+
+			// TODO: This is the same as sip_daemon.c line 372. Move to a save_bad_actor function
+			if (config->syslog_mode) {
+				syslog(LOG_NOTICE,
+				       "Source IP: %s, Method: %s, Agent: %s\n",
+				       bad_actor_event->source_ip,
+				       bad_actor_event->method,
+				       bad_actor_event->user_agent);
+			}
+
+			if (config->json_log_mode &&
+			    (json_log_bad_actor(config, bad_actor_event) !=
+			     EXIT_SUCCESS)) {
+				fprintf(stderr,
+					"Saving bad_actor json to %s failed.\n",
+					config->json_log_file);
+			}
+
+			if (db_insert_bad_actor(bad_actor_event, config) !=
+			    EXIT_SUCCESS) {
+				fprintf(stderr,
+					"Saving bad actor to db failed\n");
+			}
+
+			json_decref(json);
+			free(received_value_str);
+			bad_actor_destroy(&bad_actor_event);
+			return true;
+		} else {
+			if (config->debug_mode || config->verbose_mode) {
+				fprintf(stderr,
+					"bad_actor event_uuid already exists, not saving: %s\n",
+					event_uuid_str);
+			}
+			json_decref(json);
+			free(received_value_str);
+			return true;
+		}
 	}
 	return true;
 }
@@ -72,10 +295,17 @@ int peer_to_peer_dht_run(sentrypeer_config *config)
 		fprintf(stderr, "Starting peer to peer DHT mode...\n");
 	}
 
+	// https://github.com/savoirfairelinux/opendht/issues/590#issuecomment-1063158916
+	dht_runner_config dht_config;
+	dht_runner_config_default(&dht_config);
+
+	dht_config.peer_discovery = true; // Look for other peers on the network
+	dht_config.peer_publish = true; // Publish our own peer info
+
 	dht_runner *runner = dht_runner_new();
 	assert(runner);
 
-	dht_runner_run(runner, DHT_PORT);
+	dht_runner_run_config(runner, DHT_PORT, &dht_config);
 
 	config->dht_node = runner;
 
@@ -128,14 +358,13 @@ int peer_to_peer_dht_save(sentrypeer_config const *config,
 	ctx->runner = config->dht_node;
 	ctx->config = config;
 
-	dht_runner *runner = config->dht_node;
-	assert(runner);
+	assert(config->dht_node);
 
 	char *bad_actor_json = bad_actor_to_json(config, bad_actor_event);
 
 	dht_value *val = dht_value_new_from_string(bad_actor_json);
 	if (val) {
-		dht_runner_put(runner, config->dht_info_hash, val,
+		dht_runner_put(config->dht_node, config->dht_info_hash, val,
 			       dht_done_callback, ctx, false);
 		dht_value_unref(val);
 		free(bad_actor_json);